+++ tests/test_cacheable_pdf.py
@@ -6,7 +6,7 @@ import zfit.z.numpy as znp
 import tensorflow as tf
 
 from zfit import supports, z
-from zfit.models.cache import CacheablePDF
+from zfit.models.cache import CachedPDF
 from zfit.util import ztyping
 from zfit.util.exception import AnalyticGradientNotAvailable
 
@@ -34,7 +34,7 @@ def test_cached_pdf_equals_pdf_without_cache():
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     obs = zfit.Space("x", limits=[-5.0, 5.0])
     gauss = zfit.pdf.Gauss(mu=mu, sigma=sigma, obs=obs)
-    cached_gauss = CacheablePDF(gauss)
+    cached_gauss = CachedPDF(gauss)
     x = znp.linspace(-5, 5, 500)
 
     assert tf.math.reduce_all(tf.equal(gauss.pdf(x), cached_gauss.pdf(x)))
@@ -45,7 +45,7 @@ def test_pdf_cache_is_used():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     x = znp.linspace(-5, 5, 500)
     assert tf.equal(test_pdf.pdf_call_counter, tf.Variable(0))
     cached_test_pdf.pdf(x)
@@ -59,7 +59,7 @@ def test_pdf_cache_revaluation_if_mu_was_changed():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     x = znp.linspace(-5, 5, 500)
     assert tf.equal(test_pdf.pdf_call_counter, tf.Variable(0))
     cached_test_pdf.pdf(x)
@@ -76,7 +76,7 @@ def test_pdf_cache_revaluation_if_sigma_was_changed():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     x = znp.linspace(-5, 5, 500)
     assert tf.equal(test_pdf.pdf_call_counter, tf.Variable(0))
     cached_test_pdf.pdf(x)
@@ -93,7 +93,7 @@ def test_pdf_cache_revaluation_if_x_was_changed():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     x = znp.linspace(-5, 5, 500)
     assert tf.equal(test_pdf.pdf_call_counter, tf.Variable(0))
     cached_test_pdf.pdf(x)
@@ -110,7 +110,7 @@ def test_cached_integrate_equals_integrate_without_cache():
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     obs = zfit.Space("x", limits=[-5.0, 5.0])
     gauss = zfit.pdf.Gauss(mu=mu, sigma=sigma, obs=obs)
-    cached_gauss = CacheablePDF(gauss)
+    cached_gauss = CachedPDF(gauss)
 
     assert tf.math.reduce_all(
         tf.equal(gauss.integrate(limits=obs), cached_gauss.integrate(limits=obs))
@@ -122,7 +122,7 @@ def test_integrate_cache_is_used():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(0))
     cached_test_pdf.integrate(limits=obs)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(1))
@@ -135,7 +135,7 @@ def test_integrate_cache_is_revaluation_if_mu_was_changed():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(0))
     cached_test_pdf.integrate(limits=obs)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(1))
@@ -151,7 +151,7 @@ def test_integrate_cache_is_revaluation_if_sigma_was_changed():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(0))
     cached_test_pdf.integrate(limits=obs)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(1))
@@ -167,7 +167,7 @@ def test_integrate_cache_is_revaluation_if_limits_is_different():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(0))
     cached_test_pdf.integrate(limits=obs)
     assert tf.equal(test_pdf.integrate_call_counter, tf.Variable(1))
@@ -183,7 +183,7 @@ def test_gradient_cached_pdf():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf)
+    cached_test_pdf = CachedPDF(test_pdf)
     x = znp.linspace(-5, 5, 500)
     with tf.GradientTape(watch_accessed_variables=True, persistent=True) as tape:
 
@@ -198,12 +198,12 @@ def test_minimize_cached_pdf():
     mu = zfit.Parameter("mu", 1.0, -5, 5)
     sigma = zfit.Parameter("sigma", 1, 0, 10)
     test_pdf1 = TestPDF(obs=obs, mu=mu, sigma=sigma)
-    cached_test_pdf = CacheablePDF(test_pdf1)
+    cached_test_pdf = CachedPDF(test_pdf1)
 
     mu2 = zfit.Parameter("mu2", 5.0, -5, 5)
     sigma2 = zfit.Parameter("sigma2", 3, 0, 10)
     test_pdf2 = TestPDF(obs=obs, mu=mu2, sigma=sigma2)
-    cached_test_pdf2 = CacheablePDF(test_pdf2)
+    cached_test_pdf2 = CachedPDF(test_pdf2)
 
     testpdf = zfit.pdf.SumPDF([cached_test_pdf, cached_test_pdf2], fracs=0.5)