+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ Develop
 Major Features and Improvements
 -------------------------------
 - add ``GeneralizedCB` PDF which is similar to the ``DoubleCB`` PDF but with different standard deviations for the left and right side.
-- Added functor for PDF caching `CacheablePDF`: `pdf`, `integrate` PDF methods can be cacheable now
+- Added functor for PDF caching `CachedPDF`: `pdf`, `integrate` PDF methods can be cacheable now
 - Python 3.12 support
 - allow to use pandas DataFrame as input where zfit Data objects are expected
 - Parameter behavior has changed, multiple parameters with the same name can now coexist!